Renamed almost everything to better names
