+++ app/src/main/java/org/fontys/trackmyprint/adapters/ProductPhaseListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.tmp.trackmyprint;
+package org.fontys.trackmyprint.adapters;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -8,6 +8,8 @@ import android.widget.ArrayAdapter;
 import android.widget.ImageButton;
 import android.widget.TextView;
 
+import org.fontys.trackmyprint.MainActivity;
+import org.fontys.trackmyprint.R;
 import org.fontys.trackmyprint.database.entities.Phase;
 
 import java.util.List;
@@ -16,9 +18,9 @@ import java.util.List;
  * Created by fhict on 08/12/2016.
  */
 
-public class production_proccess_list_adapter extends ArrayAdapter<Phase> {
+public class ProductPhaseListAdapter extends ArrayAdapter<Phase> {
 
-    public production_proccess_list_adapter(Context context, List<Phase> phases) {
+    public ProductPhaseListAdapter(Context context, List<Phase> phases) {
         super(context, R.layout.production_proccess_item, phases);
     }
 
@@ -37,18 +39,18 @@ public class production_proccess_list_adapter extends ArrayAdapter<Phase> {
             public void onClick(View v) {
                 if (MainActivity.getInstance().getCurrentPhase() != pp) {
                     MainActivity.getInstance().setCurrentPhase(pp);
-                    checkIn.setImageResource(R.drawable.icon_user_vink);
+                    checkIn.setImageResource(R.drawable.productphase_checked);
                 } else {
                     MainActivity.getInstance().setCurrentPhase(null);
-                    checkIn.setImageResource(R.drawable.checkinbtn);
+                    checkIn.setImageResource(R.drawable.checkin_btn);
                 }
             }
 
         });
         if (MainActivity.getInstance().getCurrentPhase() != pp) {
-            checkIn.setImageResource(R.drawable.checkinbtn);
+            checkIn.setImageResource(R.drawable.checkin_btn);
         } else {
-            checkIn.setImageResource(R.drawable.icon_user_vink);
+            checkIn.setImageResource(R.drawable.productphase_checked);
         }
         return cv;
     }
