+++ app/src/main/java/org/fontys/trackmyprint/lib/ProductPhase.java
@@ -1,4 +1,4 @@
-package com.example.tmp.trackmyprint;
+package org.fontys.trackmyprint.lib;
 
 import org.fontys.trackmyprint.database.entities.Phase;
 
@@ -6,13 +6,13 @@ import org.fontys.trackmyprint.database.entities.Phase;
  * Created by fhict on 08/12/2016.
  */
 
-public class production_proccess {
+public class ProductPhase {
 
     public String name;
     public String iconName;
     public Phase phase;
 
-    public production_proccess(String name, String iconName, Phase phase) {
+    public ProductPhase(String name, String iconName, Phase phase) {
         this.phase = phase;
         this.name = name;
         this.name = iconName;
