+++ app/src/main/res/drawable/checkedin_btn.png
