+++ app/src/main/res/drawable/cutting_oval.png
