+++ app/src/main/res/drawable/packing_oval.png
