+++ app/src/main/res/drawable/productphase_checked.png
