+++ app/src/main/res/drawable/quality_control_oval.png
