+++ app/src/main/res/drawable/shipped_oval.png
