+++ app/src/main/res/layout/activity_map.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_map"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.tmp.trackmyprint.mapActivity">
+    tools:context=".MapActivity">
 
 
     <ImageView
@@ -22,7 +22,7 @@
     <ImageView
         android:layout_width="50dp"
         android:layout_height="50dp"
-        app:srcCompat="@drawable/profileimage"
+        app:srcCompat="@drawable/dummy_profileimage"
         android:id="@+id/profile_image"
         app:layout_constraintBottom_toBottomOf="@+id/top"
         app:layout_constraintTop_toTopOf="@+id/top"
@@ -77,7 +77,7 @@
         android:layout_marginTop="20dp" />
 
     <ImageButton
-        app:srcCompat="@drawable/printer_solid"
+        app:srcCompat="@drawable/printer_oval"
         android:id="@+id/printBtn"
         android:background="@android:color/transparent"
         android:minWidth="0dp"
@@ -95,7 +95,7 @@
         app:layout_constraintVertical_bias="0.76" />
 
     <ImageButton
-        app:srcCompat="@drawable/cutting_solid"
+        app:srcCompat="@drawable/cutting_oval"
         android:id="@+id/cutBtn"
         android:background="@android:color/transparent"
         android:minWidth="0dp"
@@ -111,7 +111,7 @@
         app:layout_constraintHorizontal_bias="0.44" />
 
     <ImageButton
-        app:srcCompat="@drawable/quality_control_solid"
+        app:srcCompat="@drawable/quality_control_oval"
         android:id="@+id/qualityBtn"
         android:background="@android:color/transparent"
         android:minWidth="0dp"
@@ -127,7 +127,7 @@
         app:layout_constraintVertical_bias="0.45" />
 
     <ImageButton
-        app:srcCompat="@drawable/packing_solid"
+        app:srcCompat="@drawable/packing_oval"
         android:id="@+id/packagingBtn"
         android:background="@android:color/transparent"
         android:minWidth="0dp"
@@ -143,7 +143,7 @@
         app:layout_constraintVertical_bias="0.26" />
 
     <ImageButton
-        app:srcCompat="@drawable/shipped_solid"
+        app:srcCompat="@drawable/shipped_oval"
         android:id="@+id/shippingBtn"
         android:background="@android:color/transparent"
         android:minWidth="0dp"
@@ -161,7 +161,7 @@
         app:layout_constraintVertical_bias="0.060000002" />
 
     <ImageButton
-        app:srcCompat="@drawable/view_on_map"
+        app:srcCompat="@drawable/view_on_map_btn"
         android:id="@+id/viewInlistBtn"
         android:minHeight="0dp"
         android:minWidth="0dp"
@@ -183,7 +183,7 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintHorizontal_bias="1.0"
-        android:background="@drawable/shipbtn"
+        android:background="@drawable/ship_btn"
         android:textColor="@android:color/background_light" />
 
 </android.support.constraint.ConstraintLayout>
