+++ app/src/main/res/layout/activity_nfc.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_nfc"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.tmp.trackmyprint.nfcActivity">
+    tools:context=".NFCActivity">
 
 
     <ImageView
