+++ app/src/main/res/layout/activity_print_details.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.example.tmp.trackmyprint.MainActivity">
+    tools:context=".MainActivity">
 
     <ImageView
         android:layout_width="0dp"
@@ -25,7 +25,7 @@
     <ImageView
         android:layout_width="50dp"
         android:layout_height="50dp"
-        app:srcCompat="@drawable/profileimage"
+        app:srcCompat="@drawable/dummy_profileimage"
         android:id="@+id/profile_image"
         app:layout_constraintBottom_toBottomOf="@+id/top"
         app:layout_constraintLeft_toLeftOf="@+id/top"
@@ -99,7 +99,7 @@
             android:layout_height="match_parent">
 
             <ImageView
-                app:srcCompat="@drawable/printer_solid"
+                app:srcCompat="@drawable/printer_oval"
                 android:id="@+id/imageView7"
                 app:layout_constraintTop_toTopOf="parent"
                 app:layout_constraintLeft_toRightOf="@+id/imageView6"
