--- app/src/main/res/layout/activity_user_process_list.xml
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context="com.example.tmp.trackmyprint.MainActivity">
-
-    <ImageView
-        android:layout_width="0dp"
-        app:srcCompat="@color/colorProfileRectangle"
-        android:id="@+id/top"
-        android:layout_height="175dp"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        android:layout_margin="0dp"
-        app:layout_constraintHorizontal_bias="0.0" />
-
-    <ImageView
-        android:layout_width="50dp"
-        android:layout_height="50dp"
-        app:srcCompat="@drawable/profileimage"
-        android:id="@+id/profile_image"
-        app:layout_constraintBottom_toBottomOf="@+id/top"
-        app:layout_constraintTop_toTopOf="@+id/top"
-        app:layout_constraintLeft_toLeftOf="@+id/top"
-        android:layout_marginStart="96dp"
-        android:layout_marginLeft="96dp"
-        app:layout_constraintVertical_bias="0.25" />
-
-    <TextView
-        android:text="Farshad Rasa"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/profile_name"
-        android:layout_marginStart="15dp"
-        app:layout_constraintLeft_toRightOf="@+id/profile_image"
-        android:layout_marginLeft="15dp"
-        app:layout_constraintTop_toTopOf="@+id/profile_image"
-        android:textColor="@android:color/background_light"
-        app:layout_constraintRight_toRightOf="@+id/top"
-        app:layout_constraintHorizontal_bias="0.0" />
-
-    <TextView
-        android:text="Last check at 5.15 pm"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/last_check_in"
-        app:layout_constraintBottom_toBottomOf="@+id/profile_image"
-        android:textColor="@android:color/background_light"
-        app:layout_constraintLeft_toLeftOf="@+id/profile_name"
-        app:layout_constraintRight_toRightOf="@+id/top"
-        app:layout_constraintHorizontal_bias="0.0" />
-
-    <ImageView
-        android:layout_width="248dp"
-        android:layout_height="33dp"
-        app:srcCompat="@drawable/checkin_status"
-        android:id="@+id/check_in_status"
-        android:layout_marginTop="32dp"
-        app:layout_constraintTop_toBottomOf="@+id/profile_image"
-        app:layout_constraintLeft_toLeftOf="@+id/top"
-        app:layout_constraintRight_toRightOf="@+id/top" />
-
-    <ListView
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:id="@+id/production_proccess1"
-        android:layout_marginStart="16dp"
-        app:layout_constraintLeft_toLeftOf="parent"
-        android:layout_marginLeft="16dp"
-        android:layout_marginEnd="16dp"
-        app:layout_constraintRight_toRightOf="parent"
-        android:layout_marginRight="16dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.18"
-        android:layout_marginTop="8dp"
-        app:layout_constraintTop_toBottomOf="@+id/top"
-        android:divider="@null"
-        android:dividerHeight="20dp"
-        android:scrollbars="none" />
-
-</android.support.constraint.ConstraintLayout>
