--- app/src/main/res/layout/user_process_item.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-    <android.support.v7.widget.CardView
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_height="75dp"
-        android:layout_marginStart="0dp"
-        app:layout_constraintLeft_toLeftOf="parent"
-        android:layout_marginLeft="0dp"
-        android:layout_marginEnd="0dp"
-        app:layout_constraintRight_toRightOf="parent"
-        android:layout_marginRight="0dp"
-        android:layout_marginTop="0dp"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_marginBottom="0dp"
-    android:layout_width="match_parent">
-
-    <android.support.constraint.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="75dp">
-
-        <TextView
-            android:text="Printing"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/production_proccess_title"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            android:paddingStart="0dp"
-            app:layout_constraintVertical_bias="0.55"
-            app:layout_constraintRight_toLeftOf="@+id/imageButtonCheck"
-            android:layout_marginEnd="8dp"
-            android:layout_marginRight="8dp"
-            android:layout_marginStart="8dp"
-            app:layout_constraintLeft_toRightOf="@+id/imageView2"
-            android:layout_marginLeft="8dp" />
-
-        <ImageButton
-            app:srcCompat="@drawable/icon_user_vink"
-            android:id="@+id/imageButtonCheck"
-            app:layout_constraintTop_toTopOf="parent"
-            android:layout_marginEnd="16dp"
-            app:layout_constraintRight_toRightOf="parent"
-            android:layout_marginRight="16dp"
-            android:layout_width="78dp"
-            android:layout_height="38dp"
-            android:layout_marginTop="24dp"
-            android:scaleType="centerCrop"
-            android:padding="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:layout_marginBottom="16dp"
-            android:background="@android:color/transparent" />
-
-        <ImageView
-            android:layout_width="56dp"
-            android:layout_height="46dp"
-            app:srcCompat="@drawable/icon_1"
-            android:id="@+id/imageView2"
-            android:layout_marginStart="16dp"
-            app:layout_constraintLeft_toLeftOf="parent"
-            android:layout_marginLeft="16dp"
-            android:layout_marginTop="16dp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:layout_marginBottom="16dp" />
-
-    </android.support.constraint.ConstraintLayout>
-</android.support.v7.widget.CardView>
